Update any.py
Still simple and ugly, but preparing for more control. I've also realized I may be doing something wrong with pywinauto as it takes away the users mouse for a brief second  to invoke clicks. Which isn't a big deal if you run this overnight or when you're not using the computer but annoying when you want this running in the background. Will try to find a way with pywinauto. Recent update makes initial progressbar detection slow, gotta see what changed or find a direct route.  Also..  I'm running this on it's trial version and I was able to crack the software to get around VM detection but I wont be be able to share instructions or the exported x64dgb patch. Just a note saying it can be done.